cambio las urls y agrego el parametro de sistema operativo en la base de datos
